+++ dev/main.ts
@@ -1,3 +1,7 @@
 import gameControl, {XBoxButton} from '../src/index.js';
 
-gameControl.on('connect', (gamepad) => {});
+gameControl.on('connect', (gamepad) => {
+	gamepad.before(XBoxButton.B, () => {
+		console.log('boom');
+	});
+});