+++ src/index.ts
@@ -7,3 +7,5 @@ import gameControl from './gamecontrol.js';
 export {gameControl};
 export {Button, XBoxButton} from './gamepad-button-maps.js';
 export default gameControl;
+
+export * from './states.js';